MacGui: use libhb built-in presets and validation functions. Update the format and save the presets in UserPresets.json, the old presets are automatically imported if the new presets file is not found. The mac gui now requires 10.7 or later.
git-svn-id: svn://svn.handbrake.fr/HandBrake/trunk@7181 b64f7644-9d1e-0410-96f1-a4d463321fa5

+++ macosx/HBAudioTrackPreset.h
@@ -24,7 +24,7 @@
 @property (nonatomic, readwrite) int sampleRate;
 @property (nonatomic, readwrite) int bitRate;
 
-@property (nonatomic, readwrite) int gain;
+@property (nonatomic, readwrite) double gain;
 @property (nonatomic, readwrite) double drc;
 
 /**
